projects/taskmasters/snapshot/docx-feature-extraction/requirements.txt
boto3
click
python-docx
pandas
nltk
transformers
bertopic